Feature ID 50: Common RPC Interface for the second RPC connection
Created a folder "RPCInterfaces". It is created for Interfaces which RPC
Endpoints should implement

Implemented a common RPC Interface for the second RPC Connection, to
force the Developers to use the right Method Signatures on frontend and
backend.

Moved RequestHandler into the Request Class for better overview
