+++ raspberry-uc-system/RaspberryUserControlSystem/CommonFiles/RPCInterfaces/IEventReceiver.cs
@@ -1,0 +1,7 @@
+namespace CommonFiles.RPCInterfaces
+{
+    interface IEventReceiver
+    {
+        void testCall(string parameter);
+    }
+}
