+++ raspberry-uc-system/RaspberryUserControlSystem/RaspberryBackend/RaspberryBackend.csproj
@@ -103,9 +103,6 @@
     <Compile Include="..\CommonFiles\Networking\ObjConn.cs">
       <Link>CommonFiles\Networking\ObjConn.cs</Link>
     </Compile>
-    <Compile Include="..\CommonFiles\Networking\others.cs">
-      <Link>CommonFiles\Networking\others.cs</Link>
-    </Compile>
     <Compile Include="..\CommonFiles\Networking\Serializer.cs">
       <Link>CommonFiles\Networking\Serializer.cs</Link>
     </Compile>
@@ -114,6 +111,9 @@
     </Compile>
     <Compile Include="..\CommonFiles\Networking\TCPServer.cs">
       <Link>CommonFiles\Networking\TCPServer.cs</Link>
+    </Compile>
+    <Compile Include="..\CommonFiles\RPCInterfaces\IEventReceiver.cs">
+      <Link>CommonFiles\RPCInterfaces\IEventReceiver.cs</Link>
     </Compile>
     <Compile Include="..\CommonFiles\TransferObjects\Request.cs">
       <Link>CommonFiles\TransferObjects\Request.cs</Link>
@@ -196,6 +196,7 @@
       <HintPath>..\packages\ABElectronics_Win10IOT\ABElectronics_Win10IOT_Libraries.dll</HintPath>
     </Reference>
   </ItemGroup>
+  <ItemGroup />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
